+++ internal/database/interfaces/sub_save.go
@@ -12,29 +12,29 @@ type SubSaveRepository interface {
 	Create(ctx context.Context, config *sub.SaveConfig) error
 
 	// GetByID 根据ID获取保存配置
-	GetByID(ctx context.Context, id int64) (*sub.SaveConfig, error)
+	GetByID(ctx context.Context, id uint16) (*sub.SaveConfig, error)
 
 	// Update 更新保存配置
 	Update(ctx context.Context, config *sub.SaveConfig) error
 
 	// Delete 删除保存配置
-	Delete(ctx context.Context, id int64) error
+	Delete(ctx context.Context, id uint16) error
 
 	// 根据任务ID获取保存配置列表
-	GetByTaskID(ctx context.Context, taskID int64) (*[]sub.SaveConfig, error)
+	GetByTaskID(ctx context.Context, taskID uint16) (*[]sub.SaveConfig, error)
 
 	// 添加保存配置与任务的关联
-	AddTaskRelation(ctx context.Context, saveID, taskID int64) error
+	AddTaskRelation(ctx context.Context, saveID, taskID uint16) error
 
 	// 添加保存配置与输出模板的关联
-	AddOutputTemplateRelation(ctx context.Context, saveID, templateID int64) error
+	AddOutputTemplateRelation(ctx context.Context, saveID, templateID uint16) error
 
 	// 添加保存配置与过滤配置的关联
-	AddFilterConfigRelation(ctx context.Context, saveID, configID int64) error
+	AddFilterConfigRelation(ctx context.Context, saveID, configID uint16) error
 
 	// 添加保存配置与订阅的关联
-	AddSubRelation(ctx context.Context, saveID, subID int64) error
+	AddSubRelation(ctx context.Context, saveID, subID uint16) error
 
 	// 添加保存配置与存储配置的关联
-	AddStorageConfigRelation(ctx context.Context, saveID, configID int64) error
+	AddStorageConfigRelation(ctx context.Context, saveID, configID uint16) error
 }
