+++ internal/models/common/base.go
@@ -4,7 +4,7 @@ import "time"
 
 // BaseDbModel 基础模型，包含所有实体的公共字段
 type BaseDbModel struct {
-	ID          int64     `db:"id" json:"id"`
+	ID          uint16    `db:"id" json:"id"`
 	Enable      bool      `db:"enable" json:"enable"`
 	Name        string    `db:"name" json:"name"`
 	Description string    `db:"description" json:"description"`
@@ -18,7 +18,7 @@ type BaseRequestModel struct {
 	Description string `json:"description"`
 }
 type BaseUpdateRequestModel struct {
-	ID          int64  `json:"id"`
+	ID          uint16 `json:"id"`
 	Enable      *bool  `json:"enable"`
 	Name        string `json:"name"`
 	Description string `json:"description"`
